+++ service-projects/src/main/java/org/opendevstack/apiservice/serviceproject/service/GenerateProjectKeyService.java
@@ -1,0 +1,11 @@
+package org.opendevstack.apiservice.serviceproject.service;
+
+import org.opendevstack.apiservice.serviceproject.exception.ProjectKeyGenerationException;
+
+public interface GenerateProjectKeyService {
+
+    String DEFAULT_PROJECT_KEY_PATTERN = "SS%06d";
+    
+    String generateProjectKey(String projectKeyPattern) throws ProjectKeyGenerationException;
+}
+