+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Created **New API module** for managing EDP Projects with create and retrieve endpoints.
+- Created **New module** external service projects to manage EDP Projects.
+
 ### External Service Jira (`external-service-jira`)
 - **New module** for checking project existance in Jira (Server)
 - Caching for the client